+++ packages/dev/next.config.js
@@ -2,6 +2,10 @@
 const nextConfig = {
   reactStrictMode: true,
   poweredByHeader: false,
+  serverRuntimeConfig: {
+    mongod: require("./server/mongo"),
+    crond: require("./server/cron"),
+  },
 };
 
 module.exports = nextConfig;