+++ packages/dev/src/components/Layout.tsx
@@ -5,6 +5,7 @@ import { useRouter } from "next/router";
 import { Logo } from "./Logo";
 import { Slash } from "@taskless/ui";
 import cx from "classnames";
+import { useInterval, useIsMounted } from "usehooks-ts";
 
 const topNav = [
   {
@@ -33,6 +34,16 @@ const useNavigationLinks = () => {
   return links;
 };
 
+const useHeartbeat = () => {
+  const isMounted = useIsMounted();
+  useInterval(() => {
+    if (!isMounted()) {
+      return;
+    }
+    fetch("/api/worker");
+  }, 10000);
+};
+
 interface LayoutProps {
   title: string;
   header?: JSX.Element;
@@ -45,6 +56,7 @@ export const Layout: React.FC<PropsWithChildren<LayoutProps>> = ({
   headerClassName,
   children,
 }) => {
+  useHeartbeat();
   const links = useNavigationLinks();
   return (
     <>